refactor: consolidate visualization to shared module for both apps
- Both apps now use shared/components/visualization.py for scatter plot
- Shared visualization has all features: zoom/pan, density modes, configurable bins
- Dynamic tooltip building works for any data columns
- Added data_version tracking for selection validation
- Moved embed_explore's render_image_preview to separate file
- App-specific visualization.py files now re-export from shared
