+++ apps/embed_explore/app.py
@@ -8,8 +8,9 @@ cluster them, and explore the results visually.
 import streamlit as st
 
 from apps.embed_explore.components.sidebar import render_clustering_sidebar
-from apps.embed_explore.components.visualization import render_scatter_plot, render_image_preview
+from apps.embed_explore.components.image_preview import render_image_preview
 from apps.embed_explore.components.summary import render_clustering_summary
+from shared.components.visualization import render_scatter_plot
 
 
 def main():
